duneserver/requirements.txt
coloredlogs
Flask
psycopg2-binary
flask-socketio
eventlet
